requirements.txt
python >= 3.8.10
numpy >= 1.20.3
scipy == 1.10.0
umap-learn >= 0.5.2
pandas == 1.3.5
matplotlib >= 3.5.0
scikit-learn >= 1.0.2
statsmodels >= 0.13.2

